api/requirements.txt
flask==2.0.2
Flask-SQLAlchemy==2.5.1
flask-migrate
psycopg2-binary
pytest